offscreen-canvas/text/2d.text.font.parse.system.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.font.parse.system</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.text.font.parse.system</h1>
<p class="desc">System fonts must be computed to explicit values</p>


<script>
promise_test(function(t) {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.font = 'message-box';
_assertDifferent(ctx.font, 'message-box', "ctx.font", "'message-box'");

return Promise.resolve();
}, "System fonts must be computed to explicit values");
</script>
